refactor(identity-core): ActorContextHolder, ActorFactory 위치 변경
