+++ bootstrap/resource-application-starter/src/main/java/com/seatliberator/seatliberator/bootstrap/security/ActorContextBindingFilter.java
@@ -1,7 +1,7 @@
 package com.seatliberator.seatliberator.bootstrap.security;
 
-import com.seatliberator.seatliberator.identity.client.actor.ActorContextHolder;
 import com.seatliberator.seatliberator.identity.core.actor.Actor;
+import com.seatliberator.seatliberator.identity.core.actor.ActorContextHolder;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
