+++ identity/identity-client/src/test/java/com/seatliberator/seatliberator/identity/client/actor/ThreadLocalActorContextHolderTest.java
@@ -1,8 +1,8 @@
 package com.seatliberator.seatliberator.identity.client.actor;
 
-import com.seatliberator.seatliberator.identity.client.actor.exception.UnidentifiableActorException;
 import com.seatliberator.seatliberator.identity.core.actor.Actor;
 import com.seatliberator.seatliberator.identity.core.actor.SimpleActor;
+import com.seatliberator.seatliberator.identity.core.actor.exception.UnidentifiableActorException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -16,7 +16,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 public class ThreadLocalActorContextHolderTest {
     private static final String MDC_KEY = "actor_context";
-    private final ActorContextHolder holder = new ThreadLocalActorContextHolder();
+    private final com.seatliberator.seatliberator.identity.core.actor.ActorContextHolder holder = new ThreadLocalActorContextHolder();
 
     @AfterEach
     void runClear() {
