+++ identity/identity-core/src/main/java/com/seatliberator/seatliberator/identity/core/actor/exception/UnidentifiableActorException.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.identity.client.actor.exception;
+package com.seatliberator.seatliberator.identity.core.actor.exception;
 
 public class UnidentifiableActorException extends RuntimeException {
     public UnidentifiableActorException(String message) {
