+++ reservation/reservation-web-mvc/src/main/java/com/seatliberator/seatliberator/reservation/infrastructure/web/book/controller/ReservationQueryController.java
@@ -1,6 +1,6 @@
 package com.seatliberator.seatliberator.reservation.infrastructure.web.book.controller;
 
-import com.seatliberator.seatliberator.identity.client.actor.ActorContextHolder;
+import com.seatliberator.seatliberator.identity.core.actor.ActorContextHolder;
 import com.seatliberator.seatliberator.reservation.application.booking.port.in.FindMyReservationUseCase;
 import com.seatliberator.seatliberator.reservation.application.booking.port.in.query.FindMyReservationQuery;
 import com.seatliberator.seatliberator.reservation.application.booking.port.in.result.ReservationResult;
