+++ reservation/reservation-web-mvc/src/main/java/com/seatliberator/seatliberator/reservation/infrastructure/web/waitlist/controller/WaitlistController.java
@@ -1,6 +1,6 @@
 package com.seatliberator.seatliberator.reservation.infrastructure.web.waitlist.controller;
 
-import com.seatliberator.seatliberator.identity.client.actor.ActorContextHolder;
+import com.seatliberator.seatliberator.identity.core.actor.ActorContextHolder;
 import com.seatliberator.seatliberator.reservation.application.waitlist.port.in.CancelWaitlistUseCase;
 import com.seatliberator.seatliberator.reservation.application.waitlist.port.in.CreateWaitlistUseCase;
 import com.seatliberator.seatliberator.reservation.application.waitlist.port.in.command.CancelWaitlistCommand;
